Add Getting Started section to README
I've added a "Getting Started" section to the READMe. It was based on the "Getting Started"-section from Code Jam 5, but I've made some small changes, like explicitly mentioning and linking the rules in this section. This makes sure that we don't have to maintain two lists of rules (one in the README and one on our website).