Fixes the broken messaging feature.
This was broken when the time ran out, but the team wants it to be
part of this repo so we can show off this feature on the stream.

As a compromise, we're not going to consider this feature when
evaluating this teams overall performance. We will evaluate this
project as if the messaging feature does not exist.